refactor(cli): fold live-persist lifecycle into _runs.py
Move all of _persist.py (setup_agent_persist, teardown_persist,
teardown_agent_persist, teardown_orchestration_persist, resolve_run_reason,
_teardown_common, _resolve_project) into _runs.py. Every importer
(agent.py, orchestrate/_orchestration.py, tests/) is rewired to
lionagi.cli._runs. _persist.py deleted.
