+++ lionagi/cli/orchestrate/_orchestration.py
@@ -24,9 +24,14 @@ from lionagi.state import provenance as _provenance
 
 from .._agents import AgentProfile, list_agents, load_agent_profile
 from .._logging import hint
-from .._persist import _resolve_project, teardown_persist
 from .._providers import build_imodel_from_spec, parse_model_spec, resolve_persisted_effort
-from .._runs import RunDir, allocate_run, save_last_branch_pointer
+from .._runs import (
+    RunDir,
+    _resolve_project,
+    allocate_run,
+    save_last_branch_pointer,
+    teardown_persist,
+)
 
 __all__ = (
     "OrchestrationEnv",
