+++ tests/apps_studio_server/test_signals_sse.py
@@ -460,12 +460,12 @@ async def test_setup_agent_persist_wires_signal_bind(tmp_path, monkeypatch):
             super().__init__(path if path is not None else db_path)
 
     monkeypatch.setattr(db_mod, "StateDB", _PatchedStateDB)
-    # Also patch the import inside _persist.py (function-local import).
-    import lionagi.cli._persist as persist_mod
+    # Also patch the import inside _runs.py (function-local import).
+    import lionagi.cli._runs as persist_mod
 
     monkeypatch.setattr(persist_mod, "StateDB", _PatchedStateDB, raising=False)
 
-    from lionagi.cli._persist import setup_agent_persist, teardown_persist
+    from lionagi.cli._runs import setup_agent_persist, teardown_persist
     from lionagi.session.signal import RunStart
 
     try:
