+++ tests/cli/test_agent_artifact_contract.py
@@ -11,8 +11,8 @@ from pathlib import Path
 import pytest
 
 from lionagi import Branch
-from lionagi.cli._persist import setup_agent_persist as _setup_live_persist
-from lionagi.cli._persist import teardown_agent_persist as _teardown_live_persist
+from lionagi.cli._runs import setup_agent_persist as _setup_live_persist
+from lionagi.cli._runs import teardown_agent_persist as _teardown_live_persist
 from lionagi.state.db import StateDB
