+++ tests/cli/test_sigint_reason_code.py
@@ -10,7 +10,7 @@ import pytest
 
 def test_resolve_run_reason_aborted_returns_cancelled_sigint():
     """Status "aborted" (KeyboardInterrupt / SIGINT) → CANCELLED_SIGINT."""
-    from lionagi.cli._persist import resolve_run_reason as _resolve_run_reason
+    from lionagi.cli._runs import resolve_run_reason as _resolve_run_reason
     from lionagi.state.reasons import RunReasons
 
     code, summary, evidence = _resolve_run_reason(status="aborted", exception=None)
@@ -22,7 +22,7 @@ def test_resolve_run_reason_aborted_returns_cancelled_sigint():
 
 def test_resolve_run_reason_aborted_not_aborted_user():
     """ABORTED_USER must NOT be written for SIGINT — it is reserved for other user aborts."""
-    from lionagi.cli._persist import resolve_run_reason as _resolve_run_reason
+    from lionagi.cli._runs import resolve_run_reason as _resolve_run_reason
     from lionagi.state.reasons import RunReasons
 
     code, _, _ = _resolve_run_reason(status="aborted", exception=None)
@@ -41,7 +41,7 @@ def test_resolve_run_reason_cancelled_sigint_is_defined():
 
 def test_resolve_run_reason_other_statuses_unaffected():
     """Non-aborted statuses must still resolve to their existing reason codes."""
-    from lionagi.cli._persist import resolve_run_reason as _resolve_run_reason
+    from lionagi.cli._runs import resolve_run_reason as _resolve_run_reason
     from lionagi.state.reasons import RunReasons
 
     code, _, _ = _resolve_run_reason(status="completed", exception=None)
